python: Split into modules
Release python library and command.

- [x] GUI tested
- [x] CLI tested
- [ ] Package library for pypi
- [ ] Package cli, gui for pypi Linux
- [ ] Package GUI for Windows
